gcp/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
